+++ packages/narrate/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/narrate
 
+## 0.58.0-pre.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.58.0-pre.0
+  - @glissade/core@0.58.0-pre.0
+
 ## 0.57.1
 
 ### Patch Changes