+++ packages/sfx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @glissade/sfx
 
+## 0.58.0-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.58.0-pre.0
+
 ## 0.57.1
 
 ### Patch Changes